+++ include/bmx/clip_writer/ClipWriter.h
@@ -80,7 +80,7 @@ public:
 
     int64_t GetDuration();
 
-    uint32_t GetNumTracks() const { return mTracks.size(); }
+    uint32_t GetNumTracks() const { return (uint32_t)mTracks.size(); }
     ClipWriterTrack* GetTrack(uint32_t track_index);
 
 public:
